requirements.txt
langchain
langchain-openai
pandas 
langchain-pinecone
python-dotenv
unicorn
fastapi